+++ include/pajlada/settings/setting.hpp
@@ -376,7 +376,7 @@ public:
     }
 
     // Returns true if the current value is the same as the default value
-    // boost::any cannot be properly compared
+    // std::any cannot be properly compared
     bool
     isDefaultValue() const
     {
@@ -406,7 +406,7 @@ private:
 
     // These are mutable because they can be modified from the "getValue" function
     mutable std::mutex valueMutex;
-    mutable OptionalType<Type> value;
+    mutable std::optional<Type> value;
     mutable int updateIteration = -1;
 
 public: