+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Breaking: Remove move ctor/operator for SettingListener. (#32)
 - Breaking: Updated to C++20. (#42)
+- Breaking: Remove boost::any `userData` support. (#44)
 - Dev: Remove `using namespace std` usages. (#38)
 
 ## v0.1.0