+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Breaking: Remove move ctor/operator for SettingListener. (#32)
 - Breaking: Updated to C++20. (#42)
 - Breaking: Remove boost::any `userData` support. (#44)
+- Breaking: Remove support for `boost::any`. (#46)
+- Minor: Added support for `std::any`. (#46)
 - Bugfix: Fixed an issue where settings without a value would always try to unmarshal the internal JSON. (#40)
 - Dev: Remove `using namespace std` usages. (#38)