+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Breaking: Remove boost::any `userData` support. (#44)
 - Breaking: Remove support for `boost::any`. (#46)
 - Breaking: Remove support for `boost::filesystem`. (#47)
+- Breaking: Remove support for `boost::optional`. (#48)
 - Minor: Added support for `std::any`. (#46)
 - Bugfix: Fixed an issue where settings without a value would always try to unmarshal the internal JSON. (#40)
 - Dev: Remove `using namespace std` usages. (#38)